# Changed out lodash functions for vanilla, improved performance of vic…