Just a little bit of investigation into speeding this tool up a bit, found the lodash utilities were likely not required. 

Not sure if this is helpful but I'd thought I'd share my findings! This was tested on a 4032 × 3024 image with 9 colours:

Pre changes

function | time
-- | --
smooth | 566ms
getLabelLocs | 14029ms
outline | 5ms
total | **14601ms**


Post changes:

function | time
-- | --
smooth | 362ms
getLabelLocs | 2203ms
outline | 5ms
total | **2572ms**

The main speed increase is from `getLabelDocs`, about an 85% increase