@daniel-munro 2025-04-22
Wow this is awesome, thanks for figuring that out and testing! Note that the pbnify.com site is now owned by Paintable Pictures, so it will be up to them to incorporate this change there.