@adamDilger 2025-04-22
Thanks! All good, I'm in-between jobs at the moment and looking for things to tinker with, I might even try moving some of the processing into WASM as an experiment, to see if there's any more performance gain to be had there 👍